refactor(clients): rewrite and reorganize existing API clients (#376)
* refactor(clients): rewrite and reorganize existing API clients

The new design provides better abstraction to decouple as much as possible
from wallabag.

A significant difference in behavior is how an incremental sync is performed.
Instead of iterating over the entries with their full data, the algorithm
computes a delta between local and server states and downloads only what is
needed. This should significantly reduce network bandwidth usage and improve
responsiveness.

This was also a good opportunity to adopt better module organization with a
src/ namespace and a barrel file.

* switch from the old to the new client interface

* remove the old code and adapt what remains

* it was a little too soon for that

* fix xcode build

* rename WStorage to LocalStorage

* add a check to ensure server compatibility

* fix endpoint paths formatting

* make server detection logic more robust